Refactored the semantic tokens file view provider to make it possible (hopefully simple) to take advantage of its features in custom LSP integrations. I've also updated the API docs to document how and why that would happen.
