+++ src/main/java/com/redhat/devtools/lsp4ij/features/semanticTokens/viewProvider/LSPSemanticTokensSingleRootFileViewProvider.java
@@ -1,0 +1,125 @@
+/*******************************************************************************
+ * Copyright (c) 2025 Red Hat, Inc.
+ * Distributed under license by Red Hat, Inc. All rights reserved.
+ * This program is made available under the terms of the
+ * Eclipse Public License v2.0 which accompanies this distribution,
+ * and is available at http://www.eclipse.org/legal/epl-v20.html
+ *
+ * Contributors:
+ * Red Hat, Inc. - initial API and implementation
+ ******************************************************************************/
+
+package com.redhat.devtools.lsp4ij.features.semanticTokens.viewProvider;
+
+import com.intellij.lang.Language;
+import com.intellij.openapi.util.TextRange;
+import com.intellij.openapi.vfs.VirtualFile;
+import com.intellij.psi.PsiElement;
+import com.intellij.psi.PsiManager;
+import com.intellij.psi.SingleRootFileViewProvider;
+import com.intellij.util.ThreeState;
+import org.jetbrains.annotations.NotNull;
+import org.jetbrains.annotations.Nullable;
+
+import java.util.List;
+
+/**
+ * Abstract base class for custom implementations of {@link LSPSemanticTokensFileViewProvider} where PSI files are
+ * generally based on {@link SingleRootFileViewProvider}.
+ */
+public abstract class LSPSemanticTokensSingleRootFileViewProvider
+        extends SingleRootFileViewProvider
+        implements LSPSemanticTokensFileViewProvider {
+
+    private final LSPSemanticTokensFileViewProviderHelper helper;
+
+    /**
+     * Creates a new semantic tokens-based file view provider for a file with a specific language.
+     *
+     * @param manager            the PSI manager
+     * @param virtualFile        the virtual file
+     * @param eventSystemEnabled whether or not the event system is enabled
+     * @param language           the file's language
+     */
+    protected LSPSemanticTokensSingleRootFileViewProvider(@NotNull PsiManager manager,
+                                                          @NotNull VirtualFile virtualFile,
+                                                          boolean eventSystemEnabled,
+                                                          @NotNull Language language) {
+        super(manager, virtualFile, eventSystemEnabled, language);
+        this.helper = new LSPSemanticTokensFileViewProviderHelper(this);
+    }
+
+    /**
+     * Creates a new semantic tokens-based file view provider for a file without a specific language.
+     *
+     * @param manager            the PSI manager
+     * @param virtualFile        the virtual file
+     * @param eventSystemEnabled whether or not the event system is enabled
+     */
+    protected LSPSemanticTokensSingleRootFileViewProvider(@NotNull PsiManager manager,
+                                                          @NotNull VirtualFile virtualFile,
+                                                          boolean eventSystemEnabled) {
+        super(manager, virtualFile, eventSystemEnabled);
+        this.helper = new LSPSemanticTokensFileViewProviderHelper(this);
+    }
+
+    public boolean isEnabled() {
+        return helper.isEnabled();
+    }
+
+    @Override
+    public boolean isDeclaration(int offset) {
+        return helper.isDeclaration(offset);
+    }
+
+    @Override
+    @NotNull
+    public ThreeState isType(int offset) {
+        return helper.isType(offset);
+    }
+
+    @Override
+    public boolean isReference(int offset) {
+        return helper.isReference(offset);
+    }
+
+    @Override
+    public boolean isStringLiteral(int offset) {
+        return helper.isStringLiteral(offset);
+    }
+
+    @Override
+    public boolean isComment(int offset) {
+        return helper.isComment(offset);
+    }
+
+    @Override
+    @Nullable
+    public String getElementDescription(@NotNull PsiElement element, @Nullable PsiElement referenceElement) {
+        return helper.getElementDescription(element, referenceElement);
+    }
+
+    @Override
+    @Nullable
+    public TextRange getSemanticTokenTextRange(int offset) {
+        return helper.getSemanticTokenTextRange(offset);
+    }
+
+    @Override
+    public void addSemanticToken(@NotNull TextRange textRange,
+                                 @Nullable String tokenType,
+                                 @Nullable List<String> tokenModifiers) {
+        helper.addSemanticToken(textRange, tokenType, tokenModifiers);
+    }
+
+    /**
+     * Returns the semantic token for the offset.
+     *
+     * @param offset the offset
+     * @return the semantic token or null if no semantic token exists at the offset
+     */
+    @Nullable
+    LSPSemanticToken getSemanticToken(int offset) {
+        return helper.getSemanticToken(offset);
+    }
+}
